(shmif+) break/bump, adding preroll stage
This is a big/important change that will require a recompile of
both arcan and everything that relies on shmif (Qemu/SDL backends etc.)

The underlying issue that more and more "possible" initial state that
needs to be negotiated have been accumulated over the years, State like
DEVICEHINT, FONTHINT, OUTPUTHINT, GEOHINT, and so on.

These events typically incur changes to rendering and output that would
cascade to resize/redraw or even more visible and costly artifacts and
delays.

To combat this, a 'preroll' stage is added underneath _open and in
the event handling routines in the lua layer. This means that the
external connection procedure (where this have been the most visible)
now roughly goes like this:

connect -> authenticate -> identify(register) -> activate ->
 (normal loop)

between register and activate, the _open call will (not included in
this change) maintain a small event-loop, absorbing the various hints
and applying to the primary segment before returning.

Some information, like initial dimensions, can thus be set immediately
(so no more _open -> _resize requirement) and there will be (next push)
a function to retrieve the activation state. This means that language
and fonts will be available "immediately" to the caller and we can cut
down on the many arguments set to tui/terminal.

On the Lua callback level, there is also a 'preroll' event added to
the normal callback handler that will be activated as soon as the
frameserver is configured. When this event has been handled (or
ignored) by the script.

This opens up a window to 'queue' all the little hints that may depend
on the type of a connection and work around the visible TOCTU bugs that
have been around in the past.