(shmif+) second round of preroll
The event is now correctly triggered, the _open call will run
until activation and arcan sends out correct activation. What
remains is to introduce this change to durden and prio, write
a test-frameserver that dumps initial state and update LWA to
also acknowledge initial state. When the tui branch is merged
the last font- and sizing changes can be applied.