(shmif) extended _control layout (req.full recomp)
Hopefully the last break for a while. Reserved areas for dealing
with memory mapped mouse cursor state synch. Also added the opt.
to carve out a custom pad area during initialization that can be
used for adding subprotocols. The main target for this is the HMD
support where it will be possible to plug in many different
latency-sensitive sensors to an avatar that we don't want on the
main queue.

FossilOrigin-Name: 17a1b4cf9b432ad77f9767228dafd0dd893d55617ca50ac802b984f5124229be