(agp) moved all GL use to be dynamic
In preparation for multi-GPU support, we need to evaluate the
possibility of running incompatible-GL vendor libraries side by side,
e.g. one MESA and one Nvidia. To get to that point, the first step is to
indirect the GL use and be able to re-initialize the table of function
pointers ourselves. This allows us to remove the direct linking
requirement for a GL library, and move that to a runtime decision.

This commit deals away with the grunt-work for the normally used
functions (though not the extensions so far) and a GL header is still
needed for the constants.

FossilOrigin-Name: d3def06c8a2b7ba3359546b315e105e2af2d4c8b3c2ca7e70ef8215fd4dc03aa