(lua) allow open_nonblock to create BCHUNK_IN/OUT
BCHUNKs are used to indicate descriptor+metadata transfer for
file- input or output. Currently, they are not completely hooked
up to the lua layer. Two things are missing : dynamic stream in/out
and sending from a specific namespace. The latter case requires
new functions to be added, and the first case is fixed with this
push.