(core++) shmif-ext vidp- buffering mode
This push is as a way to offload some of the GPU upload cost to the
client side rather than on the server when shmif is used, particularly
for cases where normal shm- use requires repacking large buffers.

Among the downsides is increased memory consumption, GPU reliance
and notably different synch/stall behavior that has not been validated
on a wider range of GPUs. Bindings that expose the feature should really
has it as an opt-in rather than a fast default at this stage.

This feature is also enabled for TUI (terminal, ...) though you need to
build both with SHMIF_TUI_ACCEL=ON and set the ARCAN_ARG=accel=1 env
in order to use it. It also removes the possibility for dirty updates
and transfers will currently always be synchronous.