(core+) initial HMD support boilerplate
Most of the pipeline work for supporting VR has been in place for
a long time, i.e. render-to-texture with a separate composition path.

This patch starts adding the boilerplate and metadata needed for
having an external tool that uses shmif- to communicat VR/HMD device
control and sensor fusion.

It needs to be complemented with a server-side fleshing out of the
stubs, and a tools/hmdbridge implementation of the producer-side.

The initial plan is to get one working for the VIVE/PSVR test env.
we have here and, at a later stage, one that wraps the needs of
OSVR, OpenVR or OpenHMD depending on which one is the least *****
to work with.