(shmif/encode) regression in encode from preroll
The preroll wait-for-activation part didn't check the segment
type to determine if a resize- was valid or not. The current state
for output segments is that _resize calls are not allowed, though
that is subject to change. This caused the encode frameserver to
wait indefinitely.

Fortunately, this also exposed an issue with the encode frameserver
giving up on receiving any event, which isn't correct.

Lastly, the apad- addition to the shmif- layout broke the circ. dep.
in calculating encode- initial size.

FossilOrigin-Name: 15a6c16712ebc52f4a16dc8f52fe431b1863cde0ab5befa20a22ed8b02b80f0a