(shmif) guard-thread semantics tune for non-auth
When running as a non-auth connection in a sandboxed environment
and different user to the server, we encounter the problem that
the guard-thread can't determine the state of its parent.
To work around that, we fall back to doing a recv/peek on the
socket to find out when it breaks.

FossilOrigin-Name: 55debe8f484e10ee35d14cf747622815f251298e3b248cc2f25ea428a4fa65d1