(lwa) fixed openal- backend patches
LWAs audio backend has been broken for quite a while due to the patches
not being in synch with upstream OpenAL refactoring of its driver
backend and this was somewhat hidden as the testing was made against
a dated git checkout.

Rewrote the patches to match the new interface, and in the process,
also switched out the cmake-external project add approach to applying
the patches as it has a few ugly fail-silent modes.