(platform/shmif) add old drmAuth approach
Troubleshooting why amdgpu et al reverts to software or fails in glamor,
along with attempts at working around the bug in Mesa where the wrong
device was picked if arcan-wayland was running on a render-node forced
the addition of the old driN approach of granting clients cardN GPU
access. This is obviously not preferred, and have to be enabled
explicitly with the ARCAN_VIDEO_ALLOW_AUTH environment.