(tools) added scratch- version of lua-tui
As part of developing the TUI API further, and providing a quicker basis for
writing arcan- compatible CLI applications, Lua- bindings and a patched
command-line tool (ltui) has been added. It is very much in pre-alpha quality
state and shouldn't be considered for any serious use for quite a while,
targetting 0.5.3.