(shmif) fix UB enum+va_args use
This theoretically breaks the ABI of the interface, but in practice not
unless you've actually enabled -fshort-enums. The reason is that the
poor enum definitions in the C- standard doesn't allow size controls for
enums other than it can't actually be larger than an int ("up to") but
the compiler is theorically allowed to pick a smaller size. This makes
it impossible to know where the variadic portion of argument transfers
start. Previously, we've protected against this by having an enum-member
that is INT_MAX in value, effectively giving the compiler no choice.
This wasn't done for the two functions patched here, on the other hand,
the VA_ARGS used was an 'abi-hack' safety-measure so little harm done.
Still it should, of course, be fixed.