(core/lua/evdev) modifers/shaders fixes
This consolidates a few related fixes. First is that the evdev
platform ignored tagging some modifiers, which adversely affected
keybindings.
Second is that decode_modifiers have been extended to be able
to provide a cheaper modifier->to_string version.
Third is the addition of allowing shader deletion. This is still
slightly incomplete/buggy for the deletion of single uniform groups,
but that will be fixed shortly.