(lua) added ALLOWINPUT flag to target_flags
With the other 'opt-in' higher capability flags added for VR, HDR,
Vector, ... it is time for the ALLOWINPUT flag. This should only be used
on a whitelisted target in the databasewhen you need to create your own
custom input events to work around bugs or limitations in the input
platform that you can't solve in some other way. Also recall that there
is a rather heavy event-queue limit, so if it is used for high-rate
devices, you need to filter on the injection side.