(build) hopefully fixed openal/arcan_lwa
A recurring problem have been the whole management of external
repositories and having to add patches to get static linking going with
the right backends. This run hopefully fixes both missing dependencies
for the first parallel make job, and pins openAL to a version where we
know the patches apply.