(core++/rebuild) favor LE pixel format
Throughout the years, the engine has consistently stuck with
GL_RGBA GL_UNSIGNED_BYTE packing as default format. With the
extended support for external clients, this results in a lot
of repacking since this format is less common than BGR on LE
sources, so we switch the channels around. This will require
a full rebuild of external/backends and supporting tools but
it's better to go through the headache now.

FossilOrigin-Name: c09923768daaefa6db71d1467e519cb7acb3f4072105fbf2dfeb757095aaa00f