(shmif) REBUILD req, fixed _event size
This was an ugly oversight that should've been detected statically,
though the test had apparently been disabled in my commit hooks.
The addition of the last_words to the shmif- page and the fsrv-
structure accidentally extended the size to 144 bytes. Changed the
structure and added a size- safeguard to keep it at 128 bytes per
event.