(platform/build/) nvidia/multi-GPU groundwork
Starting to setup a hardware environment for testing/working with the
merge of the egl-nvidia and the egl-dri platform. This needs to be done
in quite a few steps as the difference between the 'old' drm API and the
atomic one is quite large and complex, and quite annoying in our case
where we want to support both gbm- and streams- like handle mapping on
top of the new Vulkan one.

This patch works around some regressions introduced with the split to
dynamically loaded AGP a while back. It also introduces the option to
dynamically switch active AGP context (needed for multi-GPU) and to have
different function resolution schemes (also needed for multi-GPU).