(decode) add loop arg, startup slowdown fix
The decode- frameserver is still in a pretty tragic state (note, avcodec
prefers larger alignment for start than 64b). This band-aid fixes one of
the issues with startup delays from it running through the entire plugin
list of probes. It also adds a playback looping argument and fix a stall
for input during shutdown.