(platform/egl-dri) refactor to allow streams
As part of multi-GPU work and getting rid of the egl-nvidia branch, this
patch splits setup and buffer management in order to accomodate
different schemes. It also covers dynamic (re)load of EGL similarly to
how it is done for accelerated graphics in order to handle different
GPUs requiring different EGL/GL implementations.