(test) added ioinject
Added a simple test that spam-injects keyboard input. This feature is
not enabled by default on a connection point, you have to explicitly
enable it in the script on a per-connection basis as an informed
decision that you take the added risk. Typically this would be launched
as part of a target/config entry or as a one-time use connection point
in a controlled environment (preferrably a bit later when we have
asymmetric authentication).

FossilOrigin-Name: 1dd951fc605782bf84592e44b8db9ae245340101eb1510b2e3368a02d9fb0a15