(core+) gamma-bridge full-chain working
This brings us closer to the end of the first extended subprotocols,
(and by the looks of it, the harder one) that allows for bidirectional
synch of color-ramps/gamma lookup tables. The testcase is a simple appl
that enables ramp-synch sub-proto for a connection point, when a
connection comes, it sends the current LUTs. The associated test
frameserver retrieves the LUTs, increments their values and send them
back. The appl receives this event and forwards the new tables etc.

FossilOrigin-Name: 03508b46f8a7db25602a7474913e36f87d7e32da16769d75e5bf307d033e9913