(platform/egl-dri) synch.strategy work
Synchronization strategies are profiles that determines how responsive
the engine should be to requests from clients and the various subsystems
like event, audio and video. They are defined through the video layer as
that is the primary synchronization task. Since we lack a conductor and
good I/O multiplexable primitives for everything (like the terrible
POSIX semaphores) or GPU- request fencing (the GL2.1 requirement),
stronger synchronization rework will have to be postponed until we have
better driver support for more recent synchronization methods like G-Synch
and FreeSync, and a testing framework for these kinds of things. In the
meantime, we will have to make due with something simpler. This patch adds
support for a 'fast' and a 'conservative' profile, where the 'fast' profile
will keep both the event loop, video feed polling and video-tick updates
while waiting for output displays to synch, whileas the conservative will
do nothing. The default is still to maintain audio buffering and otherwise
wait between sync operations.

In the longer perspective, this approach should be able to coordinate with
CPU governors as well. With the appl-scripts taking care of user-pattern
derived idle detection, delegating suspend/resume to clients, switching
strategies and modifying other outputs like backlight, we should start to
get into a good place when it comes to mobile energy consumption.