(core+lua) custom per-segment upper size limits
When the scripting layer permits a segrequest to go through, the new
segment has the same built-in size restriction (ultimately graphics
layer and driver restrictions on texture size) as the primary segment.
This may be excessive and even dangerous for certain segment types, such
as CURSOR. In order to allow better memory and resource management, the
target_flags function is extended with the ability to set a
custom/artificial upper width/height limit during client initiated
resize operations.