(shmif) add first stub of shmif_server
This is part of the ongoing effort in strengthening the separation
between engine code, platform code and shmif code. The first round
is focused around shmif- server side code since that will open the
door for a number of important developments.

While there is some work left to get this up and running, the task
shouldn't be that heavy. The main tools to build or enchance using
this is TUI (proxy and/or share connections with subprocesses) and
for per-client network connections. Other benefits include setting
up headless/cheap continuous test and fuzzing suites that can keep
running without the volatility of being GPU dependent.