(build) allow version string to use lw-tags
It seems like 0.5.2 got tagged as a lightweight tag and not an annotated
tag. While not a big issue in itself, it also caused the --describe to
point to the wrong version when generating the versioned tag. This is
worse since it is used in crash dumps and so on to know what kind of a
build we are dealing with. Instead of messing with the tags further,
the build system is changed to accept lightweight tags in describe.