(core/shmif-server) refactoring fsrv/platform
ore work strengthening the separation between engine and platform when
it comes to managing shmif connections towards being able to run
servers- separately from the main engine. This batch reworks allocation
and management of external listening points for sockets, and splits out
the authentication/polling step to rely less on engine- related
functions. A rough estimate is that this puts the refactoring at about
halfway done.