(shmif) add support for HANDOVER subsegment
This one is a bit odd, but may well turn out important in the end.  A
longer article will be pushed later on why this is needed, but the gist
of the feature is that it is possible to request a subsegment that will
act as a primary connection (i.e. not a subsegment at all). This allows
clients that do not have access to any other connection primitives to
negotiate connections for new processes, but with a trackable ancestry.
This allows, for instance, TUI/terminal to use afsrv_decode to open up a
playback window that can take its place in window positioning/routing
scheme temporarily. When shmif-server works for proxying, we have all
needed components for full plan9-esque connection multiplexing.