(lwa) better treatment for rt- map_video_display
This should fix some issues with running durden nested in arcan-lwa.
One of the issues with the previous approach was invalid dirty tracking,
resulting in screens not updated, but also with sync_copy always being
enabled. We now check if the mapped vstore is part of a rendertarget,
and in that case enable double-buffer swapped RT updates. There might be
some gains left by timestamping RT updates and checking those before
issuing a swap.