(core+lua) extended mesh model
A big set of changes, with minor details left to work out. Extending the
featureset to cover the needs defined in glTF2.0 requires adding access
to more shader attributes (binormals, tangents, colors) and skeletal
animation attributes (bones, weights). These have been added to the mesh
model, though there is still no automatic generation of
tangents/binormals. image_tesselation has also been updated with
completed accessors.

From what I can grasp of the spec, we are still missing ways of setting
arrays of shader uniforms which will be needed for skeletal animation.
In order to link the animation system that already exists, we should
have a way to map the tree of resolved linked null_images to a
transformation-uniform-matrix-array for a model. That would allow the
animations to be specified as normal transformation chains on the
null_images in 3d space and have it neatly map unto the active skeleton.