(core) fixes #108 (glitches on VT switch)
The reason for this rather interesting bug was indeed connected to
the logical timekeeping. Since the CLOCK is taken from an epoch at
first setup, it is rebased when we return from VT suspend, but the
reference timestamp was created from the elapsed number of ticks *
tickrate and delta to epoch. This would lead to negative ticks and
that breaks animation scheduling, causing the interesting loops.

FossilOrigin-Name: 2789c5dd5925e4fb64195b0d46d79bbfb02f0e436f591b14c831b10cab8047fb