(shmif) shmif-server getting closer
Some of the shmif-server features are starting to work within the
context of the shmifsrv test application. Shouldn't be that much work
left until the basics - subsegments, autoclocking, etc. are working and
with that, proxying. Still no functions to access subfeatures like color
control, or accelerated handle passing. That will come when the
corresponding AGP/AEP support libraries have been exposed.