(core+) reworking rendertarget-density management
This is the first out of 2/3 batches that rework how the engine manages
outputs with different densities. Previously it has been up to the
scripts to size things based on an assumed default. This rework attempts
to change this by adding a configurable horiz/vertical density attribute
to each rendertarget. Vectorized assets that are to be rasterized will
be done so by the basis of the target density of the primary assets.

Subsequently, attaching/detaching such an asset from a rendertarget
should force a rerasterization to match the new format. This commit will
also rework how icons and tabs/positioning are managed by the
render_text class of functions. Future changes will cover the use of
HDR/floating point rendertargets, multisampling-FBOs and upgrades to the
format string system to allow better positioning.

Slight chance this commit will break parts of render_text that
afaik. no-one really relied upon.