(shmif/tui) bump version, change LABELHINT
The format for labelhint was quite terrible, and the parts that would
make it a useful feature didn't carry enough data (so it was in practice
never really used). This push allows the client to more accurately
expose both the labels and the binding that it suggests as the default.

FossilOrigin-Name: 57aeba78be02d073ca8561a8ade93ee5458b735b94c393d6322351b54537d7d5