(core+) more restructuring to support float+msaa
The first approach was too naive. This time around, the actual storage
format for floats are defined in alloc_surface (so that we later can
update load_image when we can deal with HDR images). Then MSAA with
a fallback to normal FBOs have been added. It's not entirely functional
as we still need to figure out where to put the blit stage, but this
should be most of it.