(tui) fix issue with force_bitmap/missing builtin
This was a slightly fun one - If we force bitmap rendering and the display
server connection provides a new font for us to use for a certain font-size
slot, the force_bitmap will try and open the font via the internal- loader
rather than ttf. This will fail, but not before we actually clear the font-slot
(since the merge- flag is not) sent and there's a dependency between the font
loader, the size and the hash-table. The result caused some font sizes to
becoming missing. The quick solution here is to auth the header of the font
before releasing the slot.