(platform/agp,gles2) repeating rPi BCM bringup
Have the need for the proprietary garbage driver to get access to
a touch display for a few projects, so here we go. This should fix
most build problems, but the idea of getting actual working FBOs
on this steaming turd is nigh impossible so likely have to rework
video_init to take more fine-grained config (needed anyhow so) and
have a render-mode that doesn't go via FBOs.