(platform/fsrv) fsrv audio fix
This one took a while to get caught because it strangely presented
no issues on my systems but broke on quite a few others. In essence,
the frameserver refactor to allow more code reuse failed to set the
local settings cache correctly for channel and samplerate, leading
to audio buffer calls with 0 channels as output targets.