(waybridge) add initial keyboard seat support
This adds the whole 'generate keymap and send descriptor' stage as well
as forwarding evdev- codes. This seems to be insufficient for
weston_terminal to actually do anything with them, even though they are
being delivered. Chances are that there's some inane XKB_ setting that
needs to be part of the environment somewhere.