(core/lua/platform) allow rt-less WORLDID
This is falling back to an old management style where we allowed WORLDID
to be processed without a rendertarget. The compromise now is that
delete_image(WORLDID) will drop the rendertarget and the vstore
(disallowing recording of the output as a downside) in order to save
memory bandwidth and allow direct- surface rendering where FBOs doesn't
exist. This is primarily to satisfy two use cases:

 1. broken platforms like BCM- raspberry Pi, FBOs are useless here
 2. special cases like egl-dri.

.2 warrant some explaining: for cases like durden, where everything goes
through script-controlled explicit rendertargets, WORLDID is never
really used and just wastes a full framebuffer. This is a prelude to the
next real optimization where we map color attachments directly to the
scanout in egl-dri rather than going through an extra blitting stage,
saving noticable amounts of bandwidth.