(arcterm) change timing / input response
This should hopefully yield a better result when it comes to the
tradeoff between inputs and refreshes. The new measurements sortof
follow the principle that if the processing is saturated, refresh anyhow
at a ~30Hz rate unless there's been user input. Otherwise, keep a 60Hz
cap on updates. The 'jumping vim cursor' symptom is still there (option
to solve it adds more latency and isn't bullet proof), since this is a
problem for 'legacy' term and not the tui- switch - this will have to do
for now.