(shmif/tui) refactor/rework tui
Hopefully last of the major changes to TUI. This patch breaks out tui in
a shape that it can eventually work as an API on its own without the
explicit dependency to arcan/shmif. There's still a lot of tedium to
work through in order to polish it into shape, but the structure here is
the likely prime candidate.