(tui/shmif) basic double-buffer scaffolding
The previous design for n-buffers in shmif was based on clients that did
full-buffer replacement and worked more like a hybrid between
triple-buffering as a strategy and classic double buffering. This mixes
poorly with dirty- regions and would require patches to the shmif-
structure that can't be done within this version. Adds basic double-
buffer (ARCAN_ARG=dblbuf) support but ones that force full-dirty.