(tui) initial shaped- renderpath support
This one is really costly at the moment, and comes with some other
legacy caveats as terminals in particularly are strongly biased towards
monospaced rendering. The approach here is an attempt at finding
something of a middle ground. In essense we add a cell attribute that
acts as a shape- reset and re-aligns the current drawing position back
to its intended cell position. This leaves it up to the caller to
indicate when there's a logical break or not. For a normal terminal,
this can usually be hacked to the set of line-drawing characters and
a set of normal |/+\ style entries but not always (due to ligatures)
Invalidation/redraw currently requires a full re-blit of the entire
row, and getting around this constraint would require quite a lot more
logic in the draw_shaped row-scan pass.

FossilOrigin-Name: 7a3d0c7da98be8908f9f3996fedf7ee69253ae2e0d60e091449b1a6eeb92c9b1