(waybridge) add stub data-device manager
Apparently the reason for GTK not requesting the seat protocol so that
we can test input turned out to be that it won't do that unless the
data-device-manager protocol is also enabled. Added that along with more
mouse event management (should not be much left for that to start
working). Now gtk3-demo actually responds to keyboard input.