(waybridge) respect dirty-updates, more input
"Some" of the GTK- mouse cursor issues was down to us not checking
the version and sending frame- events (about as undocumented as
everything else, go figure), which should be fixed now. Also start
respecting damage- regions (1 aggregate) with buffer commits. This
should be revisited when we can have n-buffering with per-buffer
damage.