(waybridge) 100% cpu fix, map more interfaces
There was a sneaky 100% CPU usage problem steming from lack of a
match between surface and arcan connection when the cursor- source
was switched between windows.

Subsurface allocation and positioner allocation dangling ptr-deref
fixed (didn't actually allocate the resource or return a proper
error), there is still one around in data-device but not as big of
a deal.