(waybridge) add quick hack around missing mods
With the design of xkbcommon and it's "integration" into wayland, it
seems like there's a handful of intractable issues when it comes to the
input model used in arcan/shmif. This can possible be hacked around on
the appl- level by fake-releasing keycodes that would contribute to
changing the modifier mask, but no solution is pretty here.