(core) fix use of custom txcos with frameset
This fixes an issue where the texture coordinates would fail to
propagate if the dst-vobj and the src-vobj already hade the same
backing store. A case that would trigger this could look like:

a = fill_surface(64, 64, 255, 0, 0)
b = null_surface(1, 1)
image_sharestorage(a, b)
image_framesetsize(b, 5)
image_set_txcos(a, {0.5, 0.2, 0.1, 0.2, 0.3, 0.4, 0.1, 0.2})
set_image_as_frame(b, a, 1)

a would be set as backing store in frameset b at slot 1, but since
backingstore in a and b already matched, the texture coordinates
would be damaged rather than copied.