(core+) chack segment_token allocation scheme
The segment_token allocation was really haphazard and only in a semi-usable
state. Reworked all the relevant interfaces to allow the caller to specifiy an
identification token that clients can use to reference other clients in
viewport hints, and switched the default engine behavior to use vid ^ cookie
for that.