(tui) rework shaping interface
There were some design issues with the previous strategy that required
some rethinking. The new approach is to allow the callback to provide
substitutiton, but leave shaping decisions to the backend. The middle-
road is that if harfbuzz etc. is built in, the _defaults call will
reference an internal callback that performs the shaping / substitution
using all the internal interface that we don't want exposed in the
normal API.