(openbsd) rough keyboard input+more bringup
At the point where durden is configurable/usable. For keyboard input,
there's still some work in regards to repeat-rate controls, keymap fields
to forward/translate, and modifier-state tracking before moving on to
mouse/trackpad input. More worrying is that we seem to crash (in stack-
smash protector no-less) during fdpass when setting up a shmif connection,
and the descriptor-passing/setup isn't enough to get new clients going.
Unfortunately, gdb works so bad as to be almost useless when compiled with
clang, and dated g++ is not an option.