(core/tui) harfbuzz shaping/substitutions
There are still some kinks to work out with the use here, so to
enable, make sure to have the harfbuzz dependency fulfilled and
run with ARCAN_ARG=substitute afsrv_terminal (or tui app.). The
big issue right now is that (testing with firacode as font)
<- << etc produce the correct glyph, but -> >> etc. do not.
Uncertain if this is due to the way the glyph renderer works
(overdraw) or if it is the way that harfbuzz input/output is
being interpreted.