(platform) added preliminary SDL2 support
This is mostly a hacked up version of the sdl1.2 platform, notably with some
event- model details that don't translate well. Though the default remains 1.2
on OSX and similar platforms, this has been enabled to easier experiment with
retina- like densities and as a path of testing both the wayland- bridge and
for running on other wayland compositors.