(core+) continued config- lookup rework
This push removes more external database context lookup in favor of an
indirection function. It also extends the config interface to allow a
fallback to a database lookup in arcan or arcan_lwa namespace (depending
on build). This should be the last step before migrating evdev/egl-dri/
other platforms to query using the config interface - which is required
for multi-GPU setup (we need to know per device libraries to load, per
device set of nodes and synch- mechanism etc). plain old getenv isn't
good enough for this, but this interface along with possible config-file
to arcan_db entries should suffice.