(core/lua) rework segment-cookie management
The strategy of using vid- as segment cookie had some risks and drawbacks,
particularly when mapped to wayland subsurfaces that it was necessary to switch
it out for something more complex.

Changed so that the segment_request call provides the parent cookie to the
script as part of the event, and the target_accept call returns the cookie
of the new segment (so that it can track reference close to usage).