(shmif) two crash-recovery bugs squashed
The first one was an omission in _RESET that changed the internal
FDs without giving the client an option to figure out which one it
was that was lost. This affected Xarcan that kept a separate polling
loop running without fetching the handle from the context each time.

The second one was much more insidious: the extended setup mode
accidentally did not set the type during acquire, leading to a missing
registration event and the connection not requiring the right type.
This didn't matter much for all settings, but one where it did matter
was the case when a crash recovery builds a new extended context using
the type of an old one. This lead to the recover- context trying to
resynch video-buffer contents from the old without having access to the
previously negotiated type. On the server side, this was treated as a
client was still being in the process of getting its type, having its
buffer transfer request being ignored.

This should possibly be complemented with a 'sanity' check that ignores
UNKNOWN segments that are locked in a buffer- transfer state and release
the semaphores as if the transfer had gone through. A symtom to look out
for is clients stuck in semaphore signalling when the shmmon- tool
checks out correctly.