(tui) adjust wrong padding calculations
There seem to be a nastier issue in play here where there is a
discrepancy between what the terminal- subgroups believe to be the
current dimensions, the tui screen management and the acon dimensions.
The workaround currently is to just force a window resize and it seems
to go away, but as the case with other advanced tui features, the ltui
tool and test scripts are now sorely needed.