(core/vrbridge/lua) complete mapping path
Dynamic limb discovery and event delivery path for the vr set of functions
should now successfully reach the scripts, enough for calls to vr_map_limb and
vr_metadata. The next step is the final poll/copy/update to the associated 3d
models and we should be set to start experimenting with actual devices.