(core/tui) flush cache after probe
This should close the gap on usable/functional ligatures.  When running in
shaped mode, there's an issue with harfbuzz returning font indices and a
limitation in the way ttf rendering works that forces us to be by index or by
value per cache flush cycle. The probe function that was being run at the
beginning to estimate cell size from font caused the cache to be polluted at
those indices. Explicitly flushing the cache after probe seems to have solved
it.