(vrbridge) invert y- axis / type confusion
the IMU data coming from HMD had the output quaternion in the y=down
format, switched for head tracking to behave properly.

Additionally, freemodel() was being called as part of DESTROY even when
the source model was a camera. Changed to properly cleanup.

Vrtest2 is still not entirely up to sniff, the scene needs more
complexity and the distortion values from OpenHMD etc.  are not correct.