(shmif) add shmifext_isext
It is sometimes useful to be able to determine the inner state of a
shmifext_setup prepared context, particularly in cases where we are
revoked handle-passing privileges or GPU- migration forcing fallback to
headless+readback- like behavior. This function can be used to probe the
current state of a shmif context.