(vrbridge/core) tuning to vr subsystem
Extended the map-limb calls to allow specification which of the
properties that are to be inherited/derived and which should come
from the device itself. This was necessary for proper IPD tuning.

Reverted/modified some of the orientation-translation for mapping
between the camera coordinate system (we usually invert y via a
scale on the model).

The vrtest2 has been extended to allow more display parameters to
be tested via the keyboard. Additional reference geometry was also
added.