(core/vr) type-confusion from vrbridge
VRbridge- performs a fair number of tricks to wrap the fact that
it's trying to act as a frameserver. This works 'mostly' but has
a type confusion in the queuetransfer- function as it didn't work
with the fsrv- structure direct.

FossilOrigin-Name: 15e1fc7cfe7308327d4bbd09f5290c42cb355af33ac54c7a9e4cc6218176aa34