(waybridge) support shm->gl transfer in bridge
A noticeable performance penality on shm- buffers comes from a forced
extra copy in that the allocation and buffer management is not
compatible between shmif and wayland. This push adds an option to enable
texture-uploads on shm surface from the bridge. These are then passed as
dma_bufs to the upstream arcan connection. Although we lose the ability
to do partial uploads (due to fake texture double buffering) the benefit
is that the server side doesn't stall or jitter due to additional
texture upload operations, and transfers can happen while upstram is
locked on VSYNC.