(waybridge) add -exec mode
Given the many 'hard to find, verify or get from documentation' edge
cases in everything wayland, we need three more modes of execution to
reduce the impact of malicious wayland clients. This push covers the
first, single -exec which build the corresponding XDG_.. dirs, runs a
single client and then cleans up / exits. This fills the case where
you only want/need to run a single wayland client every now and then.

The other two are: activating a seccomp profile, and (the difficult
due to the wayland-server design) a client-fork-out mode where the
display is used as rendezvous and the clients are then split out into
single processes of their own in order to keep the bridge running as
a daemon, but be protected from missed corner cases, leaks and all
other kinds of nastyness.