(tools/netproxy) initial skeleton
Started mapping up the network- proxy that will be used to map shmif- clients
to a networking tunnel. Although this is one of the main targets of the 0.6
branch, it will be polished as a side-task due to snail pace that some of the
lower graphics stack work.

FossilOrigin-Name: 7ca86963568f206a02296fc7a9a59df2ae6256702119a9010a828a688f7f23fa