(platform/sdl2) update mouse and keyboard
This should bring the SDL2 platform closer to being usable, at least the test
cases in durden shows it running with native retina resolution and working
keyboard mouse. Some additional hacks are probably needed to get the unicode
input to match the model use here, as that have changed a lot since SDL1.2.